myproject/requirements.txt
django
Pillow
djangorestframework
django-cors-headers
djoser
psycopg2
django-rest-knox